Refactor release state label into its own component
